backend/requirements.txt
flask
mysql-connector-python
werkzeug
python-dotenv
flask-wtf
email-validator
